Moved serialization routines to utils file
Also cleaned up imports
