+++ qmt/data/__init__.py
@@ -8,3 +8,4 @@ from .part_data import Part3DData
 from .thomas_fermi_data import ThomasFermiData
 from .schrodinger_poisson_data import SchrodingerPoissonData
 from .mobility_data import MobilityData
+from .data_utils import load_serial, store_serial, write_deserialised_file, serialised_file
